add FriendRequestService

+++ social-app/src/main/java/edu/ucsy/social/model/dto/view/CommentView.java
@@ -7,12 +7,12 @@ public class CommentView implements Serializable {
 
 	private static final long serialVersionUID = 1L;
 
-	private int id;
+	private long id;
 	private String content;
 	private String userName;
 	private LocalDateTime updatedAt;
 
-	public CommentView(int id, String content, String userName, LocalDateTime updatedAt) {
+	public CommentView(long id, String content, String userName, LocalDateTime updatedAt) {
 		super();
 		this.id = id;
 		this.content = content;
@@ -20,11 +20,11 @@ public class CommentView implements Serializable {
 		this.updatedAt = updatedAt;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
